auth-service/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
python-jose
grpcio
grpcio-tools
httpx
passlib[bcrypt]
bcrypt==4.2.0
